Api lnhpd dev to prod

+++ lnhpdWebApi/Models/ProductDose.cs
@@ -11,16 +11,16 @@ namespace lnhpdWebApi.Models
         public int dose_id { get; set; }
         public string population_type_desc { get; set; }
         public int age { get; set; }
-        public int age_minimum { get; set; }
-        public int age_maximum { get; set; }
+        public double age_minimum { get; set; }
+        public double age_maximum { get; set; }
         public string uom_type_desc_age { get; set; }
-        public int quantity_dose { get; set; }
-        public int quantity_dose_minimum { get; set; }
-        public int quantity_dose_maximum { get; set; }
+        public double quantity_dose { get; set; }
+        public double quantity_dose_minimum { get; set; }
+        public double quantity_dose_maximum { get; set; }
         public string uom_type_desc_quantity_dose { get; set; }
-        public int frequency { get; set; }
-        public int frequency_minimum { get; set; }
-        public int frequency_maximum { get; set; }
+        public double frequency { get; set; }
+        public double frequency_minimum { get; set; }
+        public double frequency_maximum { get; set; }
         public string uom_type_desc_frequency { get; set; }
     }
 }
